Move `IsPrimaryTablet` to topotools, which is probably a better shared home for it
